Create Master redirection file
This file should function as a master redirect file, though the pages I set as redirects are currently just placeholders and need to be updated with the real file name and destination address. I included two placeholders, so that you can also see that a comma is used when there is more than one redirect, but not after the last redirect on the list.