Head-Up: Copy edits and notes below this subject line
I made some heavy edit to the last topic, beginning on Line 115. I made the changes because it left a few unanswered questions for the audience or ended up too ambiguous for the audience in general. I tried to clear that up by rephrasing/rewriting content so that it provides either a solution or clears up any ambiguities around what to do or whether we're just trying to inform the audience as to what to expect about a particular problem without invoking any kind of doubt about the product, tools, development process in the minds of the audience.