Add instructions to include a PhotonView and a PhotonTransformView so that LunarModule moves on remote when user drags it
After following the current set of instruction and importing the packages downloaded from https://docs.microsoft.com/en-us/windows/mixed-reality/mrlearning-sharing(photon)-ch2,  I did not see the lunar module moving on the remote end when I moved the lunar module. This is because there is no photon component that actually syncs the position, rotation of the LunarModule across the network. This change adds steps that instruct the user to add the PhotonView and PhotonTransformView components needed to make sure that the Lunar Module's position and rotation is synchronized across users.