Speech2: Content change (affects Speech1)
Replaced ch2 'Adding an offline mode for local speech-to-text translation' with 'Using speech recognition to execute commands' which was some of the steps in the old ch1.

Current ch2 tutorial 'Adding an offline mode for local speech-to-text translation' was therefore completely removed. This because the feature was never implemented and the functionality provided with LunarcomOfflineRecognizer.cs was fake. Although this tutorial states 'translation' it's assumed this was a typo and that it should have been 'transcription', regardless, Azure speech cannot be used without internet connection and transcription is not available in MRTK by default. Consequently, it's assumed that the original intent was to implement MRTK speech commands as fallback when device is offline, however, speech commands were already covered in Getting Started tutorials series and we now provide a 'tip' section with link to these instructions.