Told user how to find the MRTK confugration menu
But note there's a problem - if this menu is not open, the options for "Audio Spatializer" dropdown  (discussed in the text below where I made the edit) will not be available for changing. Please update these documentations to tell the user how to get to that dropdown. I don't know how to do that.