Create requesting-access-to-extended-et-api
Instructions for developers how to request access to Extended ET API, we're releasing in the Iron timeframe (so we don't want this documentation to go live before April). 
Given the sensitivity of use (the API is built primarily for medical scenarios), we will pick the companies we wan to grant access to this API. 
I expect few more updates to this page next week, as we review the approval process with the business team.