Adjust trivially incorrect mentions of "Windows Mixed Reality home"
"Windows Mixed Reality home" replaced with its correct naming "Mixed Reality home" which seems to be the correct terminology? Changed a mention of "Mixed Reality" to "Windows Mixed Reality" to reflect a reference to the software instead of the platform. These adjustments may be trivial, please tell me what you think of them.