Update to error 13-14 and windows 10 versions?
In a DM with Ryan Handy (Quality Program Manager for HP VR) (u/Voodooimaxx on Reddit), I reported an issue that leads to error 13-14 on select system configurations. He told me that he would forward it to Microsoft and Nvidia (he likely misunderstood that I was testing on an AMD Radeon RX 580). This would mean this issue has been sent through developer channels, or will be in the near future. If you want to verify the validity of this bug being investigated, feel free to contact him (Microsoft probably has better channels to contact him, considering he only uses his Reddit once a month). 

Also, if anyone can do it, please make the layout consistent. Error messages are quoted in some descriptions, but left out in others, the order of the steps are shown with numbers in some descriptions, in other descriptions dots are used for steps, etc.